// web/src/generated/api/models/knowledgeSourceKind.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * BookWise AI API
 * HTTP API for BookWise AI.
 * OpenAPI spec version: 0.1.0
 */

export type KnowledgeSourceKind = typeof KnowledgeSourceKind[keyof typeof KnowledgeSourceKind];


export const KnowledgeSourceKind = {
  FAQ: 'FAQ',
  POLICY: 'POLICY',
  PREPARATION: 'PREPARATION',
  OTHER: 'OTHER',
} as const;