// web/src/generated/api/models/knowledgeSourceStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * BookWise AI API
 * HTTP API for BookWise AI.
 * OpenAPI spec version: 0.1.0
 */

export type KnowledgeSourceStatus = typeof KnowledgeSourceStatus[keyof typeof KnowledgeSourceStatus];


export const KnowledgeSourceStatus = {
  PENDING: 'PENDING',
  READY: 'READY',
  FAILED: 'FAILED',
} as const;